simplify rpc codec logic
